ignore public system  namespaces from kubernetes
