+++ pkg/kube/kubetypes/types.go
@@ -24,6 +24,7 @@ import (
 	"k8s.io/apimachinery/pkg/watch"
 
 	"istio.io/istio/pkg/cluster"
+	"istio.io/istio/pkg/slices"
 	"istio.io/istio/pkg/util/sets"
 )
 
@@ -75,6 +76,41 @@ var _ DynamicObjectFilter = staticFilter{}
 // NewStaticObjectFilter returns a DynamicObjectFilter that does not ever change (so does not need an AddHandler)
 func NewStaticObjectFilter(f func(obj any) bool) DynamicObjectFilter {
 	return staticFilter{f}
+}
+
+// composedFilter offers a way to join multiple different object filters into a single one
+type composedFilter struct {
+	// The primary filter, which has a handler. Optional
+	filter DynamicObjectFilter
+	// Secondary filters (no handler allowed)
+	extra []func(obj any) bool
+}
+
+func (f composedFilter) Filter(obj any) bool {
+	for _, filter := range f.extra {
+		if !filter(obj) {
+			return false
+		}
+	}
+	if f.filter != nil {
+		return f.filter.Filter(obj)
+	}
+	return true
+}
+
+func (f composedFilter) AddHandler(fn func(selected, deselected sets.String)) {
+	if f.filter != nil {
+		f.filter.AddHandler(fn)
+	}
+}
+
+func ComposeFilters(filter DynamicObjectFilter, extra ...func(obj any) bool) DynamicObjectFilter {
+	return composedFilter{
+		filter: filter,
+		extra: slices.FilterInPlace(extra, func(f func(obj any) bool) bool {
+			return f != nil
+		}),
+	}
 }
 
 // Filter allows filtering read operations
